Fixes issue #2291 - Move WelcomeFileManager into its own extension
