+++ extension/micro/pom.xml
@@ -78,6 +78,12 @@
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>cloud.piranha.extension</groupId>
+            <artifactId>piranha-extension-welcomefile</artifactId>
+            <version>${project.version}</version>
+            <scope>compile</scope>
+        </dependency>
         <!-- provided -->
         <dependency>
             <groupId>cloud.piranha.core</groupId>
