+++ extension/micro/src/main/java/module-info.java
@@ -41,5 +41,6 @@ module cloud.piranha.extension.micro {
     requires cloud.piranha.extension.wasp;
     requires cloud.piranha.extension.webannotations;
     requires cloud.piranha.extension.webxml;
+    requires cloud.piranha.extension.welcomefile;
     requires cloud.piranha.core.api;
 }
