+++ extension/standard/pom.xml
@@ -96,6 +96,12 @@
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>cloud.piranha.extension</groupId>
+            <artifactId>piranha-extension-welcomefile</artifactId>
+            <version>${project.version}</version>
+            <scope>compile</scope>
+        </dependency>
         <!-- provided -->
         <dependency>
             <groupId>cloud.piranha.core</groupId>
